+++ SeriesGuide/src/com/battlelancer/seriesguide/provider/SeriesGuideDatabase.java
@@ -118,12 +118,12 @@ public class SeriesGuideDatabase extends SQLiteOpenHelper {
         String LIST_ITEMS_COLUMNS = "listitem_id AS _id,list_item_id,list_id,item_type,item_ref_id";
         String LIST_ITEMS_COLUMNS_INTERNAL = "_id AS listitem_id,list_item_id,list_id,item_type,item_ref_id";
         String SHOWS_COLUMNS = LIST_ITEMS_COLUMNS
-                + ",series_id,seriestitle,overview,poster,network,airstime,airsdayofweek";
-        String SHOWS_COLUMNS_INTERNAL = "_id as series_id,seriestitle,overview,poster,network,airstime,airsdayofweek";
+                + ",series_id,seriestitle,overview,poster,network,airstime,airsdayofweek,status,nexttext,series_nextairdatetext";
+        String SHOWS_COLUMNS_INTERNAL = "_id as series_id,seriestitle,overview,poster,network,airstime,airsdayofweek,status,nexttext,series_nextairdatetext";
         String SEASONS_COLUMNS = LIST_ITEMS_COLUMNS
-                + ",series_id,seriestitle,combinednr AS overview,poster,network,airstime,airsdayofweek";
+                + ",series_id,seriestitle,combinednr AS overview,poster,network,airstime,airsdayofweek,status,nexttext,series_nextairdatetext";
         String EPISODES_COLUMNS = LIST_ITEMS_COLUMNS
-                + ",series_id,seriestitle,episodetitle AS overview,poster,network,airstime,airsdayofweek";
+                + ",series_id,seriestitle,episodetitle AS overview,poster,network,episode_firstairedms AS airstime,airsdayofweek,status,season AS nexttext,episodenumber AS series_nextairdatetext";
     }
 
     interface References {