+++ SeriesGuide/src/com/battlelancer/seriesguide/ui/CheckinActivity.java
@@ -234,7 +234,6 @@ public class CheckinActivity extends BaseActivity implements LoaderCallbacks<Cur
                         .findViewById(R.id.TextViewShowListAirtime);
                 viewHolder.poster = (ImageView) convertView.findViewById(R.id.showposter);
                 viewHolder.favorited = convertView.findViewById(R.id.favoritedLabel);
-                viewHolder.collected = convertView.findViewById(R.id.collectedLabel);
 
                 convertView.setTag(viewHolder);
             } else {
@@ -300,7 +299,6 @@ public class CheckinActivity extends BaseActivity implements LoaderCallbacks<Cur
 
         public View favorited;
 
-        public View collected;
     }
 
     interface CheckinQuery {