+++ SeriesGuide/src/com/battlelancer/seriesguide/util/Utils.java
@@ -536,7 +536,12 @@ public class Utils {
         String currentPref = prefs.getString(SeriesGuidePreferences.KEY_EPISODE_SORT_ORDER,
                 EpisodeSorting.OLDEST_FIRST.value());
 
-        return EpisodeSorting.fromValue(currentPref);
+        EpisodeSorting sorting = EpisodeSorting.fromValue(currentPref);
+        if (sorting == null) {
+            return EpisodeSorting.OLDEST_FIRST;
+        } else {
+            return EpisodeSorting.fromValue(currentPref);
+        }
     }
 
     /**
@@ -797,6 +802,9 @@ public class Utils {
         switch (theme) {
             case 1:
                 SeriesGuidePreferences.THEME = R.style.ICSBaseTheme;
+                break;
+            case 2:
+                SeriesGuidePreferences.THEME = R.style.SeriesGuideThemeLight;
                 break;
             default:
                 SeriesGuidePreferences.THEME = R.style.SeriesGuideTheme;